+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next release
 
 - support default value for `Enums` in Dart code generation.
+- removed `location` field from `ObjectBase` and its subtypes as it was not used.
 
 ## 0.2.0 (04-Dec-2023)