+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next release
+## 0.3.0 (07-Dec-2023)
 
 - support default value for `Enums` in Dart code generation.
 - removed `location` field from `ObjectBase` and its subtypes as it was not used.